Move Ephemeral registry to submodule
