+++ extensions/components/src/lib.rs
@@ -3,7 +3,7 @@ mod delivery;
 mod storage;
 mod wakeup;
 
-pub use contact_registry::EphemeralRegistry;
+pub use contact_registry::ephemeral::EphemeralRegistry;
 pub use contact_registry::http::{HttpRegistry, HttpRegistryError};
 pub use delivery::*;
 pub use storage::*;
